Consolidated Code
Removed excess functions and added into the existing loops.
Replaced excess addField() with .setDescription()

Changed named of constants and variables for consistency.

Formatted Dates in Epoch to be used by Discord's Markdown <t:> formula which shows the date in the format as chosen by the user reading it. (GMT and EU/NA formatting from the end user's settings.)